[depthMap] Rename hpp/cpp files SemiGlobalMatchingRc to Sgm
