+++ src/aliceVision/depthMap/Sgm.cpp
@@ -4,7 +4,7 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "SemiGlobalMatchingRc.hpp"
+#include "Sgm.hpp"
 #include "volumeIO.hpp"
 
 #include <aliceVision/system/Logger.hpp>
