+++ src/aliceVision/depthMap/depthMap.cpp
@@ -10,10 +10,10 @@
 #include <aliceVision/mvsUtils/MultiViewParams.hpp>
 #include <aliceVision/mvsUtils/fileIO.hpp>
 #include <aliceVision/mvsData/imageIO.hpp>
-#include <aliceVision/depthMap/SgmParams.hpp>
-#include <aliceVision/depthMap/RefineParams.hpp>
 #include <aliceVision/depthMap/Refine.hpp>
-#include <aliceVision/depthMap/SemiGlobalMatchingRc.hpp>
+#include <aliceVision/depthMap/RefineParams.hpp>
+#include <aliceVision/depthMap/Sgm.hpp>
+#include <aliceVision/depthMap/SgmParams.hpp>
 #include <aliceVision/depthMap/cuda/PlaneSweepingCuda.hpp>
 
 namespace aliceVision {
